* Added multiple changes to project

+++ admin/includes/view_all_comments.php
@@ -64,7 +64,7 @@
                 echo "<td>$comment_date</td>
                     <td><a href='comments.php?approve=$comment_id'>Approve</a></td>
                     <td><a href='comments.php?unapprove=$comment_id'>Unaprove</a></td>
-                    <td><a href='comments.php?delete=$comment_id'>Delete</a></td>
+                    <td><a onClick=\" javascript: return confirm('Are you sure you want to delete'); \" href='comments.php?delete=$comment_id'>Delete</a></td>
                 
             </tr>";
